Tweak accent colour and drawables
The accent colour resource has been renamed to "accent" (was:
"accent_1") to reflect that it's the only accent colour.

A semi-transparent version of the same colour has been added alongside,
that is used for pre-4.0 list view item backgrounds in the schedule
sessions list.

Other accent drawables were also updated to match the current accent
colour.